Move define_identity to ids.rs.
