+++ crates/sierra/src/lib.rs
@@ -8,6 +8,7 @@ pub mod compiler;
 pub mod edit_state;
 pub mod extensions;
 pub mod fmt;
+pub mod ids;
 pub mod program;
 
 lalrpop_mod!(
